aws-agent/requirements.txt
flask
python-dotenv
boto3
google-generativeai